refactor : moving pages into their proper folders
